[REFACTOR] fixture 패키지 생성
