+++ src/test/java/com/soptie/server/support/fixture/CompletedMemberDailyRoutineFixture.java
@@ -1,4 +1,4 @@
-package com.soptie.server.support;
+package com.soptie.server.support.fixture;
 
 import com.soptie.server.member.entity.Member;
 import com.soptie.server.memberRoutine.entity.daily.CompletedMemberDailyRoutine;
